refactor(model-quality): Rename model assertions to model quality (#21)
