+++ docs/docs/pages/export/filter_export.md
@@ -9,16 +9,26 @@ description: Filter data according to metrics and tags.
 
 The Filter & Export feature allows you to export data according to specific metric values or the tags created during the data exploration process.
 
-After you have created a filtering rule, click on **Generate COCO file** and sit tight, this can take up to a few minutes depending on the number of samples. 
+After you have created a filtering rule, click on **Generate COCO file** and sit tight, this can take up to a few minutes depending on the number of samples.
 
 Once the COCO file is generated, the **Download filtered data** button will unlock. Click it to download the resulting file.
 
 ![export_filter.png](../../images/export_filter.png)
 
-## Export to Encord 
+## Export to Encord
 
-Additionally, you can also export the data directly to the Encord labelling platform. For this you will need to specify dataset and project names, and optionally a description. 
+Additionally, you can also clone the data and labels directly to the Encord labelling platform.
 
-Once you are satisfied, press the **Create** button and the data will be exported to Encord.
+1. Click the _🏗 Clone_ button
+2. Choose a dataset and project title
+3. Pres the _➕ Create_ button
+
+The data will be exported to Encord. You can follow the progress in the buttom of the page.
 
 ![export_encord.png](../../images/export_filter_encord.png)
+
+:::info
+
+Every image and video selected will be cloned with all labels associated to them.
+
+:::
