--- docs/docs/pages/model-assertions/_category_.json
@@ -1,9 +1,0 @@
-{
-  "label": "Model Assertions",
-  "position": 3,
-  "collapsible": true,
-  "link": {
-    "type": "generated-index",
-    "description": "All the components that you can see on the Model Assertions page of the Encord Active app."
-  }
-}
