+++ docs/docs/workflows/improve-your-models/identify-false-negative.md
@@ -4,10 +4,11 @@ sidebar_position: 4
 
 # Identifying False Negatives
 
-Using the false negatives tab in Encord Active, you can quickly find out which objects the model misses. 
+Using the false negatives tab in Encord Active, you can quickly find out which objects the model misses.
 
 `Prerequisites:` Dataset, Labels, Predictions
-  
+
 ### Steps:
-1. Navigate to the _Model Assertions_ > _False Negatives_ tab on the left sidebar.
+
+1. Navigate to the _Model Quality_ > _False Negatives_ tab on the left sidebar.
 2. Observe the ground-truth objects (purple boxed objects) that are missed by the model and get insights on where the model fails. The remaining objects in the image are the model predictions for that image.
