+++ docs/docs/workflows/improve-your-models/understand-performance-metrics.md
@@ -13,14 +13,16 @@ Performance metrics are supported for bounding-boxes (object detection) and poly
 `Prerequisites:` Dataset, Labels, Predictions
 
 ### Steps
-1. Navigate to the _Model Assertions_ > _Metrics_ tab on the left sidebar.
+
+1. Navigate to the _Model Quality_ > _Metrics_ tab on the left sidebar.
 2. Under the **Subset selection scores**, you will see the average precision (AP) and average recall (AR) for each class in the graph to the left
-and Precision-Recall curves for each class on the graph to the right.
+   and Precision-Recall curves for each class on the graph to the right.
 3. You can select classes of interest and change IoU threshold on the upper sidebar to customize plots.
 4. On the **Mean scores** plot, you can observe in which classes the model has difficulty and in which classes it does well.
 5. According to insights you get here, you can, e.g., prioritize from which classes you need to collect more data.
 
 ### Example
+
 Comparing **person** and **clock** objects.
 
 ![clock_vs_person_performance](../../images/clock_vs_person_performance.png)
