+++ pyproject.toml
@@ -1,7 +1,7 @@
 [tool.poetry]
 name = "encord-active"
 version = "v0.1.15"
-description = "Enable users to improve machine learning models in an active learning fashion via data quality, label quality, and model assertions."
+description = "Enable users to improve machine learning models in an active learning fashion via data, label, and model quality."
 authors = ["Cord Technologies Limited <hello@encord.com>"]
 classifiers=[
     "Environment :: Console",
