+++ src/encord_active/app/model_quality/sub_pages/__init__.py
@@ -9,7 +9,7 @@ import encord_active.app.common.state as state
 from encord_active.app.common.page import Page
 
 
-class ModelAssertionsPage(Page):
+class ModelQualityPage(Page):
     @property
     @abstractmethod
     def title(self) -> str:
